+++ src/cray/cfs/operator/events/session_events.py
@@ -469,8 +469,8 @@ class CFSSessionController:
                 name='ansible-' + str(i),
                 image=self.env['CRAY_CFS_AEE_IMAGE'],
                 resources=client.V1ResourceRequirements(
-                    limits={'memory': '6Gi', 'cpu': '8'},
-                    requests={'memory': '4Gi', 'cpu': '500m'}
+                    limits=json.loads(self.env['CRAY_CFS_ANSIBLE_CONTAINER_LIMITS']),
+                    requests=json.loads(self.env['CRAY_CFS_ANSIBLE_CONTAINER_REQUESTS'])
                 ),
                 env=[
                     self._job_env['SESSION_NAME'],